initial upload
This version has been running on my desktop for months. I think there is a small bug in that if the ~/Pictures/wallpapers symlink doesn't exist or if it is a directory, the script doesn't quite know how to act. I think that if I find a wallpapers folder that is actually a folder, i need to rename it to wallpapers__ so it'll at least become part of the rotation.